+++ ssc/cmod_tcsmolten_salt.cpp
@@ -658,7 +658,7 @@ public:
 				{
 					size_t nc = steps.front().size() + 2;
 					ssc_number_t *ssc_hist = allocate("opt_history", nr, nc);
-					for (int i = 0; i<nr; i++){
+					for (size_t i = 0; i<nr; i++){
 
 						for (size_t j = 0; j<steps.front().size(); j++)
 							ssc_hist[i*nc + j] = (ssc_number_t)steps.at(i).at(j);
@@ -745,7 +745,7 @@ public:
 			util::matrix_t<double> helio_pos_temp = as_matrix("helio_positions");
 			size_t n_h_rows = helio_pos_temp.nrows();
 			ssc_number_t *p_helio_positions_in = allocate("helio_positions_in", n_h_rows, 2);
-			for (int i = 0; i < n_h_rows; i++)
+			for (size_t i = 0; i < n_h_rows; i++)
 			{
 				p_helio_positions_in[i * 2] = (ssc_number_t)helio_pos_temp(i, 0);
 				p_helio_positions_in[i * 2 + 1] = (ssc_number_t)helio_pos_temp(i, 1);
@@ -841,10 +841,10 @@ public:
                 throw spexception("The time step duration must be evenly divisible within an hour.");
         }
 
-        int n_steps_fixed = steps_per_hour * 8760;	//[-]
+        size_t n_steps_fixed = (size_t)steps_per_hour * 8760;	//[-]
         if( as_boolean("vacuum_arrays") )
         {
-            n_steps_fixed = steps_per_hour * (int)( (sim_setup.m_sim_time_end - sim_setup.m_sim_time_start)/3600. );
+            n_steps_fixed = steps_per_hour * (size_t)( (sim_setup.m_sim_time_end - sim_setup.m_sim_time_start)/3600. );
         }
         //int n_steps_fixed = (int)( (sim_setup.m_sim_time_end - sim_setup.m_sim_time_start) * steps_per_hour / 3600. ) ; 
 		sim_setup.m_report_step = 3600.0 / (double)steps_per_hour;	//[s]
@@ -1281,7 +1281,7 @@ public:
 					util::matrix_t<float> &p_udpc_T_htf_hot = allocate_matrix("ud_T_htf_ind_od_out", n_T_htf_hot_in, ncols);
 					for (int i = 0; i < n_T_htf_hot_in; i++)
 					{
-						for (int j = 0; j < ncols; j++)
+						for (size_t j = 0; j < ncols; j++)
 						{
 							p_udpc_T_htf_hot(i, j) = (float)T_htf_parametrics(i, j);
 						}
@@ -1290,7 +1290,7 @@ public:
 					util::matrix_t<float> &p_udpc_T_amb = allocate_matrix("ud_T_amb_ind_od_out", n_T_amb_in, ncols);
 					for (int i = 0; i < n_T_amb_in; i++)
 					{
-						for (int j = 0; j < ncols; j++)
+						for (size_t j = 0; j < ncols; j++)
 						{
 							p_udpc_T_amb(i, j) = (float)T_amb_parametrics(i, j);
 						}
@@ -1299,7 +1299,7 @@ public:
 					util::matrix_t<float> &p_udpc_m_dot_htf = allocate_matrix("ud_m_dot_htf_ind_od_out", n_m_dot_htf_ND_in, ncols);
 					for (int i = 0; i < n_m_dot_htf_ND_in; i++)
 					{
-						for (int j = 0; j < ncols; j++)
+						for (size_t j = 0; j < ncols; j++)
 						{
 							p_udpc_m_dot_htf(i, j) = (float)m_dot_htf_ND_parametrics(i, j);
 						}
@@ -1416,7 +1416,7 @@ public:
 
         //Load the solar field adjustment factors
         sf_adjustment_factors sf_haf(this);
-		int n_steps_full = (int)weather_reader.m_weather_data_provider->nrecords(); //steps_per_hour * 8760;
+		size_t n_steps_full = weather_reader.m_weather_data_provider->nrecords(); //steps_per_hour * 8760;
 		if (!sf_haf.setup(n_steps_full))
 			throw exec_error("tcsmolten_salt", "failed to setup sf adjustment factors: " + sf_haf.error());
         //allocate array to pass to tcs
@@ -1591,8 +1591,8 @@ public:
 				size_t n_wlim_series = 0;
 				ssc_number_t* wlim_series = as_array("wlim_series", &n_wlim_series);
 				if (n_wlim_series != n_steps_full)
-					throw exec_error("tcsmolten_salt", "Invalid net electricity generation limit series dimension. Matrix must have "+util::to_string(n_steps_full)+" rows.");
-				for (int i = 0; i < n_steps_full; i++)
+					throw exec_error("tcsmolten_salt", "Invalid net electricity generation limit series dimension. Matrix must have "+util::to_string((int)n_steps_full)+" rows.");
+				for (size_t i = 0; i < n_steps_full; i++)
 					tou.mc_dispatch_params.m_w_lim_full.at(i) = (double)wlim_series[i];
 			}
 
@@ -1998,12 +1998,12 @@ public:
 			H_rec / rec_aspect /
 			double(heliostatfield.ms_params.m_n_flux_x));
 
-		for( int i = 0; i < n_rows_eta_map; i++ )
+		for( size_t i = 0; i < n_rows_eta_map; i++ )
 		{
 			flux_maps_out[n_cols_flux_maps*i] = eta_map_out[3 * i] = (ssc_number_t)heliostatfield.ms_params.m_eta_map(i, 0);		//[deg] Solar azimuth angle
 			flux_maps_out[n_cols_flux_maps*i + 1] = eta_map_out[3 * i + 1] = (ssc_number_t)heliostatfield.ms_params.m_eta_map(i, 1);	//[deg] Solar zenith angle
 			eta_map_out[3 * i + 2] = (ssc_number_t)heliostatfield.ms_params.m_eta_map(i, 2);							//[deg] Solar field optical efficiency
-			for( int j = 2; j < n_cols_flux_maps; j++ )
+			for( size_t j = 2; j < n_cols_flux_maps; j++ )
 			{
 				flux_maps_out[n_cols_flux_maps*i + j] = (ssc_number_t)(heliostatfield.ms_params.m_flux_maps(i, j - 2)*heliostatfield.ms_params.m_eta_map(i, 2)*flux_scaling_mult);		//[kW/m^2]
 			}
